Move accordion state logic to hook
