+++ packages/core/chrome/core-chrome-browser/index.ts
@@ -57,4 +57,5 @@ export type {
   SolutionNavigationDefinition,
   SolutionNavigationDefinitions,
   EuiSideNavItemTypeEnhanced,
+  RenderAs,
 } from './src';
