+++ packages/core/chrome/core-chrome-browser/src/index.ts
@@ -56,4 +56,5 @@ export type {
   SolutionNavigationDefinition,
   SolutionNavigationDefinitions,
   EuiSideNavItemTypeEnhanced,
+  RenderAs,
 } from './project_navigation';
