+++ packages/shared-ux/chrome/navigation/src/ui/constants.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+import type { EuiThemeSize, RenderAs } from '@kbn/core-chrome-browser';
+
+export const DEFAULT_SPACE_BETWEEN_LEVEL_1_GROUPS: EuiThemeSize = 'm';
+export const DEFAULT_IS_COLLAPSED = true;
+export const DEFAULT_IS_COLLAPSIBLE = true;
+export const DEFAULT_RENDER_AS: RenderAs = 'block';
